refactor: reuse terminal control sequence handling
