+++ tool/common/common.go
@@ -28,6 +28,7 @@ import (
 	"sort"
 	"strings"
 	"unicode"
+	"unicode/utf8"
 
 	"github.com/ghodss/yaml"
 	"github.com/gravitational/trace"
@@ -348,15 +349,129 @@ func FormatUserDisplay(primary, secondary, username string) string {
 	}
 }
 
-// sanitizeDisplayValue drops control characters and trims whitespace so the
-// result is safe to print in a single-line terminal cell.
+// sanitizeDisplayValue removes terminal controls, newlines, and tabs for a
+// single-line terminal cell.
 func sanitizeDisplayValue(s string) string {
+	s = StripTerminalControlSequences(s)
+	// Preserve word boundaries while keeping the value to one terminal line.
+	return strings.Join(strings.Fields(s), " ")
+}
+
+// StripTerminalControlSequences removes terminal control sequences from s,
+// preserving printable text, newlines, and tabs.
+func StripTerminalControlSequences(s string) string {
 	var b strings.Builder
 	b.Grow(len(s))
-	for _, r := range s {
-		if !unicode.IsControl(r) {
-			b.WriteRune(r)
+
+	for i := 0; i < len(s); {
+		r, size := utf8.DecodeRuneInString(s[i:])
+		if r == utf8.RuneError && size == 1 {
+			c := s[i]
+			switch {
+			case c == '\x1b':
+				i = consumeEscapeSequence(s, i+1)
+			case c == '\x90' || c == '\x98' || c == '\x9d' || c == '\x9e' || c == '\x9f':
+				i = consumeStringControl(s, i+1)
+			case c == '\x9b':
+				i = consumeCSISequence(s, i+1)
+			case c == '\n' || c == '\t':
+				b.WriteByte(c)
+				i++
+			case c < ' ' || c == '\x7f' || (c >= '\x80' && c <= '\x9f'):
+				i++
+			default:
+				b.WriteByte(c)
+				i++
+			}
+			continue
+		}
+
+		switch r {
+		case '\x1b':
+			i = consumeEscapeSequence(s, i+size)
+			continue
+		case '\u0090', '\u0098', '\u009d', '\u009e', '\u009f':
+			i = consumeStringControl(s, i+size)
+			continue
+		case '\u009b':
+			i = consumeCSISequence(s, i+size)
+			continue
+		}
+
+		if unicode.IsControl(r) && r != '\n' && r != '\t' {
+			i += size
+			continue
+		}
+
+		b.WriteString(s[i : i+size])
+		i += size
+	}
+
+	return b.String()
+}
+
+func consumeEscapeSequence(s string, i int) int {
+	if i >= len(s) {
+		return i
+	}
+
+	switch s[i] {
+	case '[':
+		return consumeCSISequence(s, i+1)
+	case ']', 'P', 'X', '^', '_':
+		return consumeStringControl(s, i+1)
+	case '(', ')', '*', '+', '-', '.', '/', '#':
+		if i+1 < len(s) {
+			return i + 2
+		}
+		return i + 1
+	}
+
+	for i < len(s) {
+		c := s[i]
+		i++
+		if c >= '\x30' && c <= '\x7e' {
+			return i
+		}
+		if c < ' ' || c > '\x7e' {
+			return i
 		}
 	}
-	return strings.TrimSpace(b.String())
+	return i
+}
+
+func consumeCSISequence(s string, i int) int {
+	for i < len(s) {
+		c := s[i]
+		i++
+		if c >= '\x40' && c <= '\x7e' {
+			return i
+		}
+	}
+	return i
+}
+
+func consumeStringControl(s string, i int) int {
+	for i < len(s) {
+		if s[i] == '\a' {
+			return i + 1
+		}
+		if s[i] == '\x1b' && i+1 < len(s) && s[i+1] == '\\' {
+			return i + 2
+		}
+
+		r, size := utf8.DecodeRuneInString(s[i:])
+		if r == utf8.RuneError && size == 1 {
+			if s[i] == '\x9c' {
+				return i + 1
+			}
+			i++
+			continue
+		}
+		if r == '\u009c' {
+			return i + size
+		}
+		i += size
+	}
+	return i
 }
